Consolidated the repository test
