+++ test/DotNetToolkit.Repository.Integration.Test/Data/TestEntity.cs
@@ -1,7 +1,5 @@
 ﻿namespace DotNetToolkit.Repository.Integration.Test.Data
 {
-    using System;
-
     public class Customer
     {
         public int Id { get; set; }
@@ -18,24 +16,5 @@
         public string State { get; set; }
         public int CustomerId { get; set; }
         public Customer Customer { get; set; }
-    }
-
-    public class CustomerWithGuidId
-    {
-        public Guid Id { get; set; }
-        public string Name { get; set; }
-        public int Rank { get; set; }
-    }
-    
-    public class CustomerWithStringId
-    {
-        public string Id { get; set; }
-        public string Name { get; set; }
-        public int Rank { get; set; }
-    }
-
-    public class CustomerWithNoId
-    {
-        public string Name { get; set; }
     }
 }
